Refactor: 게시글 수정시 기존 데이터 적용부분 가독성을 위해 setShowInfo, setShowResInfo 함수로 분리
